seniorsatwork-matching/requirements.txt
# Job Matching System - Dependencies
# Python 3.11+

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
PyMySQL>=1.1.0

# Elasticsearch
elasticsearch>=8.12.0

# ETL & parsing
phpserialize>=1.3
beautifulsoup4>=4.12.0

# OpenAI
openai>=1.12.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
